Level Editor in Progress
New prefab "The Brick" in place with its new "BricksBehaviours" script as companion.

Welcome the prestigious "ColorPresetScriptable" that contains all presets for Bricks colors System that you are able to create via the menu item and modify them (adding many presets as you want to).

Don't touch the editor it is not quite functional and bla bla bla ... I'm currently adding the Undo to it, and later, the Moving Brick editor particularity, then I'll easily add the bonus/malus brick (thanks to Vincent), and implement everything in the game.